third_party/WebKit/LayoutTests/svg/custom/svg-length-empty-string.html

<!DOCTYPE html>
<title>Empty string length value</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<svg width="" height="">
  <rect width="100" height="100" fill="green"/>
</svg>
<script>
test(function() {
  let bounds = document.querySelector('svg').getBoundingClientRect();
  assert_greater_than(bounds.width, 0);
  assert_greater_than(bounds.height, 0);
});
</script>
